Fix ingress service name reference and bump chart version
The ingress template was using smartmetserver.name which is a static value
'smartmetserver', but the actual service name is generated using the
fullname template which includes the release name.

This caused 503 errors when the ingress tried to route to a service named
'smartmetserver' instead of the actual service name like 'smartmetservertest'.

Changes:
- Fixed ingress template: {{ $.Values.smartmetserver.name }} -> {{ include "smartmetserver.fullname" $ }}
- Bumped chart version from 1.5.0 to 1.5.1

Fixes ingress routing to backend services.